Fix bfloat16 crossing into numpy/PIL/sklearn under bf16-mixed precision
precision="bf16-mixed" means encoder/decoder outputs (mu, reconstructions,
coherence-classifier predictions) are bfloat16 tensors. Neither NumPy nor
PIL support that dtype, so every place one of these tensors crosses into
sklearn, np.argmax, or to_pil_image() needs an explicit float32 cast first
-- otherwise it raises "Got unsupported ScalarType BFloat16".

- utils/eval.py: add a _to_np() helper (detach + float + cpu + numpy) and
  route every encodings/preds tensor handed to sklearn or np.argmax
  through it (train_clf_lr_*, eval_clf_lr_*, from_preds_to_acc/ap).
- mv_vaes/mv_vae.py: .float() before each t_f.to_pil_image() call on a
  reconstruction/generation grid.
- utils/text.py: .float() before tensor_to_text()'s numpy conversion of a
  decoder's text reconstruction.

+++ mv_vaes/mv_vae.py
@@ -344,7 +344,10 @@ class MVVAE(pl.LightningModule):
                     mod_grid_m = make_grid(
                         torch.cat([mod_m, mod_rec_m], dim=0), nrow=n_samples_row
                     )
-                    mod_grid_m = t_f.to_pil_image(mod_grid_m)
+                    # bf16-mixed precision means mod_rec_m (a decoder output)
+                    # can be bfloat16; to_pil_image() goes through numpy,
+                    # which doesn't support that dtype.
+                    mod_grid_m = t_f.to_pil_image(mod_grid_m.float())
                     self.logger.log_image(
                         key="reconstructions " + key, images=[wandb.Image(mod_grid_m)]
                     )
@@ -722,7 +725,11 @@ class MVVAE(pl.LightningModule):
                 elif key == "text":
                     continue
                 imgs_grid_m = make_grid(random_gen_m, nrow=n_samples_row)
-                imgs_grid_m = t_f.to_pil_image(imgs_grid_m)
+                # random_gen_m is a pure decoder output (no float32 ground
+                # truth mixed in to force promotion) -- bfloat16 under
+                # bf16-mixed precision, which to_pil_image()'s numpy path
+                # can't handle directly.
+                imgs_grid_m = t_f.to_pil_image(imgs_grid_m.float())
                 self.logger.log_image(
                     key="random generations " + key,
                     images=[wandb.Image(imgs_grid_m)],
@@ -775,7 +782,7 @@ class MVVAE(pl.LightningModule):
                         ),
                         nrow=n_samples_row,
                     )
-                    mod_grid_m_m_tilde = t_f.to_pil_image(mod_grid_m_m_tilde)
+                    mod_grid_m_m_tilde = t_f.to_pil_image(mod_grid_m_m_tilde.float())
                     self.logger.log_image(
                         key="cond_gen_" + key + "_to_" + key_tilde,
                         images=[wandb.Image(mod_grid_m_m_tilde)],
@@ -787,7 +794,7 @@ class MVVAE(pl.LightningModule):
                         ),
                         nrow=n_samples_row,
                     )
-                    mod_grid_m_m_tilde_cov = t_f.to_pil_image(mod_grid_m_m_tilde_cov)
+                    mod_grid_m_m_tilde_cov = t_f.to_pil_image(mod_grid_m_m_tilde_cov.float())
                     self.logger.log_image(
                         key="cond_gen_cov_" + key + "_to_" + key_tilde,
                         images=[wandb.Image(mod_grid_m_m_tilde_cov)],
